live: support new `igninfo.json` ISO file
This file includes information about where the Ignition embed area is
located. This will allow Ignition embedding to work on s390x, where we
can't just use the `ignition.img` trick. For backwards compatibility, we
still look for `ignition.img` if `igninfo.json` isn't found.

Add ISO fixtures with `igninfo.json`. Include both an x86_64 version
(where it points at `ignition.img`) and an s390x one (where it points at
`cdboot.img` and `ignition.img` doesn't exist).

+++ tests/images/iso-extract-minimal-iso.sh
@@ -22,6 +22,11 @@ grepq() {
 iso=$1; shift
 iso=$(realpath "${iso}")
 
+if [[ "${iso}" == *.s390x.* ]]; then
+    echo "Skipped; minimal ISO not supported on s390x"
+    exit 0
+fi
+
 tmpd=$(mktemp -d)
 trap 'rm -rf "${tmpd}"' EXIT
 cd "${tmpd}"
